Crosswalk Table

Return a subset of the crosswalk table containing codemeta properties and matching column

crosswalk_table(from, to = NULL, full_crosswalk = NULL, trim = TRUE)

Arguments

from
the name of a column in the crosswalk table to map from.

to
the name of one or more columns in the crosswalk table to map into

full_crosswalk
Path or URL (requires internet!) of the full crosswalk table.

trim
drop rows giving properties not found in the 'from' column?

Value

a tibble containing the trimmed crosswalk table, listing property (in CodeMeta), and the corresponding terms in both from and to columns.

Examples

# \donttest{
crosswalk_table(from = "GitHub", to = c("Zenodo", "Figshare"))

#> # A tibble: 11 x 4
#>    Property            GitHub        Zenodo            Figshare   
#>    <chr>               <chr>         <chr>             <chr>      
#>  1 codeRepository      html_url      relatedLink       relatedLink
#>  2 programmingLanguage languages_url <NA>              <NA>       
#>  3 downloadUrl         archive_url   <NA>              <NA>       
#>  4 author              login         creators          <NA>       
#>  5 dateCreated         created_at    <NA>              <NA>       
#>  6 dateModified        updated_at    <NA>              <NA>       
#>  7 license             license       license           License    
#>  8 description         description   description/notes Description
#>  9 identifier          id            id                <NA>       
#> 10 name                full_name     title             Title      
#> 11 issueTracker        issues_url    <NA>              <NA>       

# }