create_codemeta

create a codemeta list object in R for further manipulation. Similar to write_codemeta, but returns an R list object rather than writing directly to a file. See examples.

create_codemeta(pkg = ".", root = ".", id = NULL,
  force_update = getOption("codemeta_force_update", TRUE), verbose = TRUE,
  ...)

Arguments

pkg
package path to package root, or package name, or description file (character), or a codemeta object (list)

root
if pkg is a codemeta object, optionally give the path to package root. Default guess is current dir.

id
identifier for the package, e.g. a DOI (or other resolvable URL)

force_update
Update guessed fields even if they are defined in an existing codemeta.json file

verbose
Whether to print messages indicating opinions e.g. when DESCRIPTION has no URL. See give_opinions.

...
additional arguments to write_json

Value

a codemeta list object

Examples

cm <- create_codemeta("jsonlite")

#> Some elements could be improved, see our opinions via give_opinions('/Library/Frameworks/R.framework/Versions/3.5/Resources/library/jsonlite/.')

cm$keywords <- list("metadata", "ropensci")